Index Products -> Category/Manufacturer Description
Add in Manufacturers
Add in call to Category Class
Standardise Filename and CSS
Installation
