+++ includes/modules/content/index_products/templates/cm_ip_category_description.php
@@ -1,0 +1,5 @@
+<div class="col-sm-<?php echo $content_width; ?> cm-ip-category-description">
+  <div class="well well-sm">
+    <?php echo $cm_description; ?>
+  </div>
+</div>
