--- includes/modules/content/index_products/templates/description.php
@@ -1,5 +1,0 @@
-<div class="col-sm-<?php echo $content_width; ?> ip-description">
-  <div class="well well-sm">
-    <?php echo $description; ?>
-  </div>
-</div>
